# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestão inteligente de fotografias e eventos caninos do PS (Photo Storage).",
        "title": "PS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte PS",
            "email": "suporte@ps.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/clients": {
            "get": {
                "description": "Retorna a lista paginada de clientes da temporada do tenant autenticado com suporte a busca e filtros",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Listar clientes da temporada",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da Temporada",
                        "name": "season_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Termo de busca textual (pessoa, cão ou foto)",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Número da página (padrão 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Itens por página (padrão 10, máximo 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/client.PaginatedClients"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Cadastra um cliente vinculado a uma temporada com cães e fotos para o tenant autenticado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Criar cliente da temporada",
                "parameters": [
                    {
                        "description": "Dados do Cliente da Temporada",
                        "name": "client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/client.SeasonClient"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/client.SeasonClient"
                        }
                    },
                    "400": {
                        "description": "Requisição inválida ou referência não pertencente ao tenant",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Tenant bloqueado ou expirado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/clients/{id}": {
            "get": {
                "description": "Retorna os detalhes de um cliente da temporada pelo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Obter cliente por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/client.SeasonClient"
                        }
                    },
                    "400": {
                        "description": "ID obrigatório",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza os dados de um cliente da temporada, cães e fotos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Atualizar cliente da temporada",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do Cliente",
                        "name": "client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/client.SeasonClient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/client.SeasonClient"
                        }
                    },
                    "400": {
                        "description": "Requisição inválida ou referência não pertencente ao tenant",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Tenant bloqueado ou expirado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Cliente não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove um cliente da temporada do tenant autenticado",
                "tags": [
                    "Clients"
                ],
                "summary": "Excluir cliente da temporada",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Excluído com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ID obrigatório",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Tenant bloqueado ou expirado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "client.Dog": {
            "type": "object",
            "properties": {
                "breed": {
                    "type": "string"
                },
                "competitions_won": {
                    "type": "integer"
                },
                "is_owner": {
                    "type": "boolean"
                },
                "judge": {
                    "type": "string"
                },
                "judges": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/client.Photo"
                    }
                },
                "won_competitions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "client.PaginatedClients": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/client.SeasonClient"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "client.Photo": {
            "type": "object",
            "properties": {
                "amount_paid": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "file_number": {
                    "type": "string"
                },
                "judges": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "payment_method": {
                    "description": "Pix, Cartão de Crédito, Cartão de Débito, Dinheiro, Não pago",
                    "type": "string"
                },
                "photographer_id": {
                    "type": "string"
                }
            }
        },
        "client.SeasonClient": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "dogs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/client.Dog"
                    }
                },
                "id": {
                    "type": "string"
                },
                "person_id": {
                    "type": "string"
                },
                "season_id": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite \"Bearer \" seguido do token JWT obtido no login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}